refactor: move resolveMainFields option to uni-builder
