+++ .changeset/dry-cups-study.md
@@ -1,0 +1,9 @@
+---
+'@rsbuild/uni-builder': patch
+'@rsbuild/webpack': patch
+'@rsbuild/document': patch
+'@rsbuild/shared': patch
+'@rsbuild/core': patch
+---
+
+refactor: move resolveMainFields option to uni-builder
