+++ packages/document/docs/en/config/options/source.mdx
@@ -49,15 +49,3 @@ import TransformImport from '@en/shared/config/source/transformImport.md';
 import PreEntry from '@en/shared/config/source/preEntry.md';
 
 <PreEntry />
-
-## source.resolveExtensionPrefix
-
-import ResolveExtensionPrefix from '@en/shared/config/source/resolveExtensionPrefix.md';
-
-<ResolveExtensionPrefix />
-
-## source.resolveMainFields
-
-import ResolveMainFields from '@en/shared/config/source/resolveMainFields.md';
-
-<ResolveMainFields />
