+++ packages/document/docs/zh/config/options/source.mdx
@@ -49,15 +49,3 @@ import TransformImport from '@zh/shared/config/source/transformImport.md';
 import PreEntry from '@zh/shared/config/source/preEntry.md';
 
 <PreEntry />
-
-## source.resolveExtensionPrefix
-
-import ResolveExtensionPrefix from '@zh/shared/config/source/resolveExtensionPrefix.md';
-
-<ResolveExtensionPrefix />
-
-## source.resolveMainFields
-
-import ResolveMainFields from '@zh/shared/config/source/resolveMainFields.md';
-
-<ResolveMainFields />
